WiP: start a Finder model
This is a start of another approach to respecting the preview_only flag when
publishing finders to Rummager as well as to the Publishing API.

In the abstract finder publisher tried on another branch, nearly all the
methods took a finder, which suggests that modelling finders as objects in
their own right is a more appropriate approach.

A bunch of other things would go along with this: the finder presenters would
need updating, and there should be services and observers to do the sending to
publishing API and Rummager - it would be much more like the rest of this
application. Is it worth it?