Current finders and Documents#index
This commit adds a list of the Finders, a Documents controller and index
and exposes the current one as a helper method. It also redirects the
User to the first one in the list of Finders. So far this only has CMA
but we will add more as the work goes on.