Derserialization of Documents from Publishing API
This commit adds a `Documents#show` action which uses an instance method
on a `Document` subclass called `from_publishing_api` to turn a
Publishing API response back into an object.

This commit also adds other attributes which come back from the
Publishing API to the Document model.

Additionally, it moves the `format_specific_metadata` method from the
Presenter so we can use this to figure out the format specific fields
from the response.